Update 1.0 goal
I didn't realize that NPM actually treats minor version bumps as breaking changes in the `0.x.y` series of versions, so users still get warned about breaking changes even when Apollo Client is not 1.0. This means we should only bump to the 1.0 version once it's important to have a stable API and we consider breaking changes a lot more carefully than we do now.