update vscode settings
According to the TSLint config the line limit is 140 characters and not 100 characters so it is super annoying to see a bunch of lines wrapping at 100 characters when they are valid giving the linter.

Moves the ruler to reflect the TSLint configuration and disables column wrapping entirely as when text does go over in development it is harder to read when wrapped.