Bumping `graphql-tag`, ignoring deprecation warnings
`graphql-tag` has deprecated the export of the `printer` export (which we use for various AST-related tasks here in `apollo-client`). when `graphql-tag@v2` comes out, we can switch to using the printer / parser from the `graphql` package, but for now we know about the deprecation warning and can safely ignore it.