Update lastResult with computed value of currentResult() (#1927)
A pseudo-race condition exists for observable queries that return
a result from currentResult() before setUpQuery() is invoked. The
condition causes a referential equality checks to fail for results
obtained before and after setUpQuery(), despite the results being truly
equal.

This commit stores results computed during currentResults() and
reconfigured queryListenerForObserver() to consult the ObservableQuery
for the lastResult, instead of maintaining a reference within a closure.