Update Apollo Optics to Apollo Engine
I struggled with this just a little so I thought I'd update the docs to reflect what I found worked. I have personally included `app.user(compress());` to the configServer options based on Apollo Engine instructions, but I'm not entirely certain it should be there, that it works and that it should be in the example. So I left it out of this document update.