feat: expose a method to refetch all observed queries (#2625)
* feat: expose a method to refetch all observed queries

In a scenario where one updates state via a third party real time provider, one can run into situations where the real time provider is unable to do so. For instance with network outages. To be able to refetch all observed queries will make sure the state is in the correct state. Reset store cannot be used for this scenario, because it rejects any query that is mid-flight.

* fix code and add more tests

* fix code and add more tests