Switch to `github.pr.user.login` since `commits[].author.login` is undefined.
As demonstrated in the Danger test[0] on apollographql/apollo-client#3444,
the `author` property is not always set.  While it could make sense to
individually look at each individual commit's "login", Git commits are not
always tied to GitHub users so it seems more relevant to look at the GitHub
login of the user (or bot!) opening the PR, when making decisions about
whether to apply bot-only policies.

[0]: https://circleci.com/gh/apollographql/apollo-client/8427