Pin `@types/graphql` to 0.12.x to workaround readonly constraints
`@types/graphql` 0.13.x now configure the graphql AST as
`readonly`. Apollo Client is updating the AST in several places,
which means we can't use the new typings yet. We're going to
revisit the approach Apollo Client uses to updated the AST,
but in the meantime we're pinning `@types/graphql` to a version
that allows AST updates.