Fix bogus OptimisticWrapperFunction import path in executeStoreQuery.d.ts.
As reported by @nevir in this comment:
https://github.com/apollographql/apollo-client/pull/3394#issuecomment-402944767

Apparently, when TypeScript generates the declaration for the
executeStoreQuery function in lib/executeStoreQuery.d.ts, it "normalizes"
the import path for the OptimisticWrapperFunction type badly wrong, as a
relative ../../.. path that goes all the way up to the root of the file
system, then back down to same directory, including several ancestor
directories (like my home directory) that are obviously not portable.

Re-exporting the OptimisticWrapperFunction type from executeStoreQuery.ts
seems to mitigate the problem, though I certainly do not claim to
understand the original behavior.