Use simple string as cache key for DepTrackingCache#depend.
The default cache key for functions wrapped with the optimism library is
an immutable tuple object that is guaranteed to be === any other tuple
with the same elements. Although this abstraction is convenient and
powerful and definitely the right choice for a default makeCacheKey
function, using a string (when possible) is somewhat faster, especially in
extremely hot code paths like this one.