Stop using InMemoryCache#transformDocument internally.
Although the results of this method were cached, it was expensive for new
query objects, because it created a deep copy of every incoming query, and
was often called multiple times for the same query.

Instead of cloning and modifying the queries themselves, we can just make
sure that the reading and writing of cache data includes any __typename
fields when the addTypename option is enabled.

Code outside InMemoryCache (e.g. the QueryManager from apollo-client)
still uses the transformDocument method, so I've left the implementation
in place, but the cache no longer uses it internally.