Apollo Client integrated local state handling
This commit provides the bulk of the new Apollo Client local
state handling capabilities. It replaces the need to use
`apollo-link-state` in a link chain for local state management,
by merging local state handling directly into the Apollo Client
core. The majority of the new local state functionality can be
found in the `LocalState` class, which is tied into the
`QueryManager` to integrate with Queries, Mutations and
Subscriptions.

Key Changes:

- Replaces the need to use `apollo-link-state`.
- `apollo-link-state` defaults have been replaced with new
  initializer functions, which are a much more capable and
  flexible way to initialize the cache.
- Initializers and local resolvers can be set through the
  `ApolloClient` constructor, or set separately using new
  public API functions.
- A local schema can be set through `ApolloClient`, which
  can then be used by external tooling (like Apollo Client
  Devtools).
- Mixing remote and local (`@client` based) data together in
  queries is supported.
- Queries can now use the `@export` directive to export the
  result of a local resolver into a variable, that can then
  be automatically passed into a subsequent query. E.g.
  `@client @export(as: "someVar")`
- Full SSR support.
- Numerous bug fixes made to address the oustanding
  `apollo-link-state` issue backlog.