Remove `getTypeDefs` / `setTypeDefs` from `ApolloClient`
We're going to avoid getters/setters for `typeDefs` for now,
and instead recommend `typeDefs` are set via the `ApolloClient`
constructor. We were originally exposing getters/setters for
integrations like Apollo Client Devtools, but we'll adjust
devtools to access the `typeDefs` in a different way. We will
likely want to revisit this decision in the future, to
accommodate the possibility that people want to add to their
local schema at different points in their application, but
we'll cross that bridge when we get there. For now keeping the
changes to the public `ApolloClient` API minimal, is the goal.