Rollup bundling changes
These changes wire up a new Rollup based bundling approach. We
no longer build and provide an ESM bundle. By providing an ESM
bundle, and pointing to it via the `package.json` `module`
property, we were inadvertently reducing the effectiveness of
build tool tree shaking / dead code elimination capabilities.
For example, since webpack was getting the ESM bundle from the
`module` property, it wasn't able to get to the fully extracted
ESM source in `./lib`, to use it for DCE.

With the changes in this commit, the `package.json` continues
to point `main` to the CJS bundle (with all Apollo Client code
in it), and now points `module` to the non-bundled ESM entry
point of `./lib/index.js`. This means we're leaving the bundling
of ESM code up to Apollo Client consumers and their build tooling,
most of which will be able to leverage the expanded ESM source
to build more efficient bundles on their side.

We'll re-visit this strategy as work on AC 3 progresses, but so
far the results have been promising.