Fix TS source links; Exclude API from link checker
The source links needed to be updated to match the new non-monorepo
project structure. I'm not sure why the API is being excluded from
link checking, but it's defintiely needed. Links like
`/api/core/#ApolloClient.resetStore` aren't working, when they
should be. But since the API was being excluded previosuly,
this commit fixes the config to exclude it again.