Adjust publish process to use "dist" as the project root
These changes include a new script that is used to prepare the
"dist" directory for publishing. This includes:

- Copying a modified package.json into "dist", after it has been
  adjusted (remove "./dist" directory prefixes, remove un-needed
  package.json properties, and remove the private flag).
- Copying the README.md into "dist".

This script is now called as part of the "predeploy" script.
The "deploy" script will now also cd into "dist" before
publishing, so only the contents in "dist" will be published to
npm (and "dist" will essentially serve as the published package
root).