Re-export `gql` from `graphql-tag`
We're going to re-export `gql`, making it part of the
`@apollo/client` bundle, to pave the way for future optimizations
(that will require `gql` being part of core).

Note: Due to the way `graphql-tag` is bundled, we have to follow
Rollup's recommendations listed at
https://rollupjs.org/guide/en/#error-name-is-not-exported-by-module
to be able to include it in the `@apollo/client` bundle.