Ensure StoreObject properties always start with field name.
When a field has multiple cached values due to being written into the
cache at different times with different arguments, it's important that we
do not throw away the association between the short field.name.value (an
identifier, as defined in the schema) and the longer storeFieldName
strings computed from the field and its arguments.

This commit ensures that field.name.value is always recoverable from the
storeFieldName, by prepending field.name.value + ":" to any storeFieldName
that does not already start with field.name.value.