Support eviction of specific entity fields.
Evicting an entire entity object is often overkill, when all you really
want is to invalidate and refetch specific fields within the entity. This
observation is especially true for the ROOT_QUERY object, which should
never be evicted in its entirety, but whose individual fields often become
stale or need to be recomputed.

A critical nuance here is that fields are evicted according to their
field.name.value, rather than a specific storeFieldName, since it doesn't
make a whole lot of sense to evict the field value associated with a
specific set of arguments. Instead, calling cache.evict(id, fieldName)
will evict *all* values for that field, regardless of the arguments.