Use field.name.value for result caching, rather than storeFieldName.
In PR #5617, we shifted from using just IDs for caching results read from
the EntityStore to using IDs plus field names, which made the result
caching system significantly more precise and less likely to trigger
unnecessary rerenders.

However, the field names we used in that PR were storeFieldName strings,
instead of the actual field names you'd find in your GraphQL schema.
Within the InMemoryCache implementation, a storeFieldName is the full
string produced by Policies#getStoreFieldName, which must begin with the
original field.name.value, but may include an additional suffix that
captures field identity based on arguments (and also, but less frequently,
directives).

For the purposes of result caching, it's important to depend on the entity
ID plus the actual field name (field.name.value), rather than the ID plus
the storeFieldName, because there can be many different storeFieldName
strings for a single field.name.value (and we don't want to waste memory
on dependency tracking), and because it's safer (more conservative) to
dirty all fields with a particular field.name.value when any field with a
matching storeFieldName is updated or evicted, since field values with the
same field.name.value often overlap in subtle ways (for example, multiple
pages of a paginated list), so it seems dangerous to encourage deleting
some but not all of them.

Perhaps more importantly, application developers should never have to
worry about the concept of a storeFieldName, since it is very much an
implementation detail of the InMemoryCache, and the precise format of
storeFieldName strings is subject to change.