Provide stable options.storage object to read functions.
If a read function needs to cache expensive computations, or do any other
sort of long-term bookkeeping, it's convenient to have a unique private
storage object that gets passed to the read function each time it is
invoked for a particular field within a particular object.

Making this work for normalized entity objects that have a string ID was
easy, but it was somewhat trickier to support non-normalized, nested
objects. The trick is to use the object itself as a key in a WeakMap (via
the KeyTrie), so the object will not be kept alive after it has been
removed from the cache. We do not need or want to continue using the same
storage object after such a change, because it is never safe to assume a
non-normalized object has the same identity as any other (!==) object in
the cache.