Eliminate WatchQueryOptions.fetchResults.
The WatchQueryOptions interface is an important part of Apollo Client,
since it's the primary configuration API for (arguably) the most important
method in the library: client.watchQuery. In other words, all of its
fields must have clear meanings, and apply in a consistent way across
everything the client does.

WatchQueryOptions.fetchResults (#1752) does not meet this standard, since
it suppresses only one of several calls to queryManager.fetchQuery (when
setting new variables in ObservableQuery.setVariables), and only when
another similarly-named-but-unrelated boolean parameter (tryFetch) happens
to be false.

Since we don't use the fetchResults option internally, and it has not been
providing a predictable, satisfying experience for external users who
might be tempted to solve a problem with { fetchResults: false }, the AC3
release seems like a good time to remove it. TypeScript should clearly
flag any remaining usage in application code.