Terminology: arguments instead of variables.
When calling cache.evict, we do not have a FieldNode, so we cannot supply
variables to be combined with the field's arguments, but must instead
provide the fully-resolved arguments object.

To make this work, the context argument of KeyArgsFunction no longer
provides context.variables, since that information is not always
available. The arguments passed as the first argument to the
KeyArgsFunction should be sufficient to compute the field identity.