Implement an offsetLimitPagination field policy function.
I made this helper function an export of the the @apollo/client/utilities
entry point so it will not be included in your application bundle unless
you import and use it.

We hope you will find it useful, but we also encourage using it just for
guidance/inspiration when writing your own custom field policy helper
functions. For example, you might want to change args.offset to
args.start, or args.limit to args.max.