Invert storage of subtype-supertype relationships in Policies class.
It's more convenient to configure possibleTypes as a map from supertypes
to arrays of subtypes, since that's how a schema introspection query
reports them.

However, we can perform policies.fragmentMatches checks much more
efficiently if we invert that structure internally, using a map from
subtypes to sets of possible supertypes.

When a fragment with type condition S is tested against an object with
__typename T, we now search upwards from T through its supertypes until we
find S (fragment matches), or the search terminates (matching fails).

We can (and did) achieve the same results by starting from S and searching
downward for T, but the branching factors tend to be larger in that
direction, so the search tends to take longer.