Never return undefined from Policies#getTypePolicy.
There's no harm in returning an empty TypePolicy object for an
unrecognized type name, so we can simplify this code by removing the
option to prevent initial policy creation.

Future commits will allow type policies to be inherited according to
possibleTypes, which means an intermediate subtype can have inherited
policies even if it does not have its own entry in typePolicies. It will
be simpler to implement that behavior if we always initialize the policy
the first time we call getTypePolicy for a particular type.