Avoid warning about clobbering scalar fields.
While it is possible to write a field merge function that handles updates
for scalar field values, we should treat scalar data as opaque by default,
without displaying "Cache data may be lost..." warnings that nudge the
developer to write an unnecessary merge function.

None: because this code is inside of a process.env.NODE_ENV block, it will
not be executed or bundled in production.

Should fix #7071.