Revert "Merge pull request #7761 from apollographql/7611-preserve-no-cache-fetchpolicy"
This reverts commit 319993eac4b5ef5fad82994c76e3e7fff6a30dcb, reversing
changes made to 6120401181612e9245873cb4735d9fe0098e70a0.

Because #7611 could be a visible change for some applications, we've
decided to relocate these changes onto the release-3.4 branch, rather than
main, so we revert the PR on main, merge main into release-3.4, then
un-revert the PR on the release-3.4 branch.