Remove unnecessary watchDep member from InMemoryCache class.
This maybeBroadcastWatch logic was introduced in #6387 to cope with the
possibility that a cache watcher might switch between reading
optimistically and non-optimistically over time, depending on the
presence/absence of optimistic layers when broadcastWatches was called.

Now that cache watchers read consistently optimistically or consistently
non-optimistically (thanks to the Stub technique introduced recently),
this trick should no longer be necessary.