Remove only non-__typename fields from ROOT_MUTATION.
I suspect this change will be slightly less disruptive than completely
removing the ROOT_MUTATION object from the cache after each mutation.

It's safe to keep __typename because it isn't sensitive information,
like the arguments of root mutation fields.

It's convenient to keep __typename because that data (and the enclosing
object) would otherwise have to be recreated the next time a mutation
writes its results into the cache.

We could use the mutation.document root fields to perform this removal
more precisely, but that's more work for no clear gain.