Stabilize the execute function from useMutation().
There are instances in Apollo Studio where the execute function from
useMutation is placed in a useEffect dependency array for some reason.
The fact that the callback can change seems to cause React-185s. This
change makes the execute function the same for the lifetime of the
component.

This also incidentally seems to fix #9111 because we’re now updating the
ref directly in the render function. I dunno anymore.